% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregate_access.R
\name{disaggregate_access}
\alias{disaggregate_access}
\title{Dissaggregate accessible - and remote or protected - areas}
\usage{
disaggregate_access(
  aoi_border,
  roads_projected,
  PAs_union,
  buffer_dist = 2500,
  aoi_prefix = "aoi_",
  destination = getwd()
)
}
\arguments{
\item{aoi_border}{a simple features (sf) polygon of the area of interest border or boundary. This should be projected to the relevant CRS of the aoi.}

\item{roads_projected}{a simple features (sf) lines of the roads in the area of interest. This should be projected to the relevant CRS of the aoi.}

\item{PAs_union}{a simple features (sf) polygon of the protected areas in the area of interest, not as a multipolygon, but as a single, unified shape. This should be projected to the relevant CRS of the aoi.}

\item{buffer_dist}{a vector of the buffer distance from the road that should be considered accessible in the projection units. The default is 2.5 km from the road (in m).}

\item{aoi_prefix}{a character string of the prefix supplied to write the resulting sf objects. The default is "aoi_".}

\item{destination}{a character string of the directory to save the results to. The default is the working directory.}
}
\value{
two shapefiles written to the destination provided.
}
\description{
Disaggregate a given area of interest (aoi), such as a national or regional project area polygon, into accessible and remote areas, according to the defined buffer distance from the roads and the protected areas. The results are saved to the directory provided, rather than loading it into the workspace.
}
