% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disaggregate_gez.R
\name{disaggregate_gez}
\alias{disaggregate_gez}
\title{Dissaggregate woody cover by Global Ecological ZOne}
\usage{
disaggregate_gez(rast, poly_gez, scaling = 1)
}
\arguments{
\item{rast}{a raster of the woody cover in the area of interest.}

\item{poly_gez}{a simple features (sf) polygon of the GEZ. This should be projected to the relevant CRS of the raster.}

\item{scaling}{a vector by which the raster data will be scaled. If the raster has the woody cover as a percentage (0-100), then scaling = 1 (default). If the raster has decimals (0-1), then scaling = 100.}
}
\value{
a data frame of the total woody area per GEZ.
}
\description{
Disaggregate the woody cover data of a given raster by the Global Ecological Zones for the same extent.
}
