% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cover.R
\name{process_cover}
\alias{process_cover}
\title{Process woody cover change data.}
\usage{
process_cover(
  cover_raster,
  loss_raster,
  gain_raster,
  remote_shape,
  aoi_shape,
  aggregation = 8,
  aoi_prefix = "aoi_",
  destination = getwd()
)
}
\arguments{
\item{cover_raster}{a raster of 2000 woody cover for the aoi.}

\item{loss_raster}{a raster of 2000-2018 woody loss for the aoi. (this should be the output of process_coverchange)}

\item{gain_raster}{a raster of 2000-2012 woody gain for the aoi. (this should be the output of process_coverchange)}

\item{remote_shape}{a (sf) polygon of the remote areas. (this should be the output of disaggregate_access)}

\item{aoi_shape}{a (sf) polygon of the area of interest}

\item{aggregation}{a vector by which the raster data will be aggregated. Default = 8.}

\item{aoi_prefix}{a character string of the prefix supplied to write the resulting raster. The default is "aoi_".}

\item{destination}{a character string of the directory to save the results to. The default is the working directory.}
}
\value{
three raster images (.tif) written to the destination provided.These consist of the 2000 total, 2018 total and 2018 accessible woody cover data.
}
\description{
Mask, reclassify, aggregate and write woody cover change data from Hansen et al.
}
