% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coverchange.R
\name{process_coverchange}
\alias{process_coverchange}
\title{Process woody cover change data.}
\usage{
process_coverchange(
  change_raster,
  aoi_shape,
  aggregation = 8,
  aoi_prefix = "aoi_",
  var_suffix = "_coverchange",
  destination = getwd()
)
}
\arguments{
\item{change_raster}{a raster of woody gain/loss for the aoi.}

\item{aoi_shape}{a (sf) polygon of the area of interest}

\item{aggregation}{a vector by which the raster data will be aggregated. Default = 8.}

\item{aoi_prefix}{a character string of the prefix supplied to write the resulting raster. The default is "aoi_".}

\item{var_suffix}{a character string of the suffix supplied to write the resulting raster.. The default is "_coverchange".}

\item{destination}{a character string of the directory to save the results to. The default is the working directory.}
}
\value{
a raster image (.tif) written to the destination provided.
}
\description{
Mask, reclassify, aggregate and write woody cover change data from Hansen et al.
}
